minor edits and QoL; added end condition for ToggleIntake --Elijah
