+++ src/main/java/frc/robot/commands/Drivetrain/ArcadeDriveCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.Drivetrain;
+package frc.robot.commands.drivetrain;
 
 import frc.robot.subsystems.Drivetrain;
 
@@ -22,6 +22,10 @@ public class ArcadeDriveCommand extends CommandBase {
     this.turnSpeed = turnSpeed;
 
     addRequirements(drivetrain);
+  }
+
+  public ArcadeDriveCommand(Drivetrain drivetrain, double forwardSpeed, double turnSpeed) {
+    this(drivetrain, () -> forwardSpeed, () -> turnSpeed);
   }
 
   // Called when the command is initially scheduled.
